Created a new tabbed theme
Created a theme with tabbed interface    features --------  - filter the
treeview #104  - corrected the slowness of the tree with a new library
(jstree) #103  - added a tabbed system to show documentation in tabs
#105    under the cover ---------------  - updated jQuery to 1.6.2  -
updated jQuery UI to 1.8.16  - fixed a bug in the splitter that made a
infinite loop  - merged the scripts in one single place (default.js) ->
easier to maintain and better for the cache  - merged the trees and
removed the accordion, fixed a lot of performance problems  - removed
the "span"s in the tree, better for big trees.  - removed treeview.js
for jstree.js    Improvements to add ------------------- - tooltips in
the tree for names larger than the width of the sidebar  - do not open
twice the same file in the tabs.