Revamped logging to make optimal use of Cilex
I have rewritten the logging section to remove all custom
logging functionality. This does mean that log files are no
longer created.

The Log class still exists (albeit now namespaced) but
only serves as a struct to hold severity levels as the rest
of the application has not notion of those.

Unit tests for the logging class have been removed as it
does not 'do' anything anymore.