phpDocumentor defaults to project:run again when no command is provided
During the move to Cilex we inadvertently broke that you auto-executed the
execution process if you provided no arguments. Instead the help listing was
provided as this is the default behaviour of Cilex.

To circumvent this solution I overrided the ArgvInput class of the Symfony2
Console Component to check if the first option in the Argv array was an
argument or an option. If there are no options in the array or if the first item
starts with `-` then we insert the project:run command name.