Refactored file handling into a separate Component structure
The previous was file handling was an improvement over the one
before but we want to extract the Fileset component out of
phpDocumentor. Another added benefit was that performance is
improved by re-using Symfony's Finder component.

After this commit must the code be moved to it's own repository
and included via Composer.