Added unit tests for the XML Exporter
The XML Exporter is not unit tested at all at the moment and could use a good
test suite. This is a little harder to test but the Argument is now fully tested.
The other unit tests are marked as incomplete as reminder that they should still
be done per the example of the Argument test