Fix for #284, Windows gives 'I/O Error: Invalid Argument' (also known as errno 22)
During a twitter discussion with @FnTmLV about issue #284 he discovered that the issue
does not occur if the URI sent to the XSLTProcessor contains an additional / after the
scheme. (thus you get `file:///C:\...` instead of `file://C:\..`).
A fix in the code has been made that when running on the Windows OS that another / is
prepended to the filename.