Changed all classes from prefixed to namespaced except phpDocumentor_Reflection_*
phpDocumentor has been slowly moving towards PHP 5.3 in the past 3 alpha's. This is
causing BC break and repeating this for until beta is undesirable. As such we have
changed ALL classes from a prefixed version to a namespaced version.

With this move we hope to achieve better interoperability with other libraries and to
heighten adoption of ourown libraries. Aside from replacing underscores with slashes
have the following changes been done:

* Added better support for namespaced plugins and moved the Core plugin to a
  namespaced version
* The Exporter classes have Exporter suffixed to the classname to prevent keyword
  collisions
* All Abstract class have their last namespace part prefixed; example:
  phpDocumentor_Parser_Abstract has become \phpDocumentor\Parser\ParserAbstract.
  This is also to prevent keyword collisions.
* Moved main classes of namespaces into the folder of the namespace to make
  separating them as component easier and match namespace schemes of other projects.
  Example: pphDocumentor_Parser resides in /src/phpDocumentor but is now renamed
  to \phpDocumentor\Parser\Parser and moved to /src/phpDocumentor/Parser

These changes will have impact on Plugin maintainers (sorry guys!) and users who
directly implement the Parser and Transformer instead of using the command line.