Written an independent installer
It has become obvious that the manual installation process costs a lot of grieve and misunderstanding.
As such an installer was made that can be downloaded without the rest of phpDocumentor and once invoked
will install or update phpDocumentor from Github. In addition to the previous will Composer be invoked and
is GraphViz and the XSL extension tested.
Once the user has ran this command he should be immediately ready to go.