#491 Limited connecting to logging via command to onetime
Each logging command probably invokes the parent command to link the system.log
and system.debug events to the output of the Symfony Components. When tasks are re-used
then the output is linked multiple times resulting in duplicated error messages.

We have added a static variable that checks whether the linking has already occurred and
returns if so.