#479 special characters cause invalid structure.xml files
When a \v, \f or other binary character is used will this cause the XML file to
be invalid. By wrapping the output of PHP-Parser in a special 'PrettyPrinter'
that calls upon the originalValue attribute which is injected by our own Lexer
specialization are these binary invocations never actually transformed.