#461: Removed classmap for phpDocumentor directory
phpDocumentor used to have classes that did not conform to PSR-0.
So in the composer.json I had added the root directory to both a
classmap directive and the PSR-0 directive to generate autoload
files. This may cause conflicts in certain situation.

To be able to remove the classmap I also had to fix a few PSR-0
inconsistencies that caused classes not to be found.