#52 initial commit of the Twig writer plugin
In this commit we have layed down the groundwork for the Twig writer and
Twig extension support for phpDocumentor. A readme is included with
some basic information but this must still be expanded.

The code currently ain't pretty but that is to be improved before promoting
this plugin to the develop branch and even into its own repository