Refactored functional tests to use Behat instead of PHPT tests.
Over time it became clear that the PHPT tests had issues with special characters,
trimming and other white space issues. This combined with the inability to test
against parts of the output (when testing against an error message I do not want
to test the entire output) and easier test the XML output provided incentive to
take a look into Behat.

Behat allows us:

* to write tests in human language, thus making it easier
* to test parts of the `structure.xml` file
* to test parts of the output
* to output a nicely readable format

Behat is not installed by default; to include it you would need to run composer's
`install` or `update` command with the `--dev` argument.

After installing behat you can run the functional tests with the following command:

```
php ./bin/behat --no-paths
```

The number of tests should be expanded to match all use-cases and the existing
cases should still be modified to include the test data instead of it being in an
external file.