#516: Reverting pear.symfony-project.org change in composer.yml and committing new lock
The change as contributed appears to break Travis and my local system. It is safe
to assume it does so on every other system since I have not yet received complaints
about the issue wherefor this fix was intended.

A new resolution needs to be found; probably by upgrading phpDocumentor's event
system to the EventDispatcher v2.