Updating composer.json to fix latest BC breaks of Composer.
Composer has recently switched the default stability for
installations to 'stable' instead of 'dev'.
Since we rely on unreleased software (such as zf2 and
Cilex) does this cause major headaches.

For this reason we have followed the advice to set the
minimum stability to dev so that all dependencies are
installed correctly.