#96: Finished making changes to support the creation of PHAR files.
realpath calls are necessary for making sure we read files on the disk
and not relative to the phar root (and thus inside the phar package).
At the same time we had to adjust the compiler to not strip comments due
to annotations.

What remains is adding this to the deployment process and uploading the
phar file to the web root