#530: target was not correctly converted to an absolute path in transform command
The transform command was not refactored yet to deal with the phar packaging. This has
been adapted and now it works like a charm.
Also the file selection process has changed and only 50% of the previously selected files
are now included and the size has dropped to 7mb instead of 15