Fixes #561: DocBlocks are now also recognized as being File-level if they precede a non-documentable element
When using namespaces it is not recommended to use the @package tag but phpDocumentor tends
to not recognize the DocBlock because the business rules do not take this into account. The business
rules have now been expanded that if the first DocComment in a file precedes a non-documentable
node than it is considered to be the File-level DocBlock.