Fixed #592: file-level DocBlock should take interfaces into account
The first DocBlock in a file is considered to be a 'file-level' DocBlock with interfaces
since the rules for determining a file-level DocBlock mention that if the first DocBlock
contains an @Package tag it is a file-level DocBlock, UNLESS it precedes a class definition.

Since interfaces and classes do not receive the @package tag in a file-level DocBlock will
the interface get the 'Default' package name as it has no DocBlock of its own.

I believe this to be in error; the same rules that apply to classes apply to interfaces as well.