Fix #602: partial configuration paths cause fatals
When a configuration file only contains a part of the path for a configuration
option (such as 'transformations' but no 'template/name' for 'transformations/template/name')
the application will crash because it tries to 'get()' the next node but the parent
is an empty string.

To fix this an additional check should be added that will verify if a child in a chain
is a valid object; if not then the default should be taken as if the option is entirely
absent.