Switch to using PHPUnit using Composer, add Mockery
PHPUnit just released version 3.7 with support for Composer. Although travis offers phpunit
out of the box we do not want to rely on an external service for our dependencies but
manage these ourselves. This will allow us to control which version to use and do controlled
upgrades.

As a PoC we also added Mockery to try out whether this works 'better' than using the out of
the box PHPUnit mocking library. Should this not satisfy our needs we can switch back.