Add behat task, update PHPUnit task and add generic pre-push task
With the addition of behat should it be added to the phing make file
so that we can use it once we resurrect Jenkins.
PHPUnit just got installed using composer and as such should the phing
task no longer use the system-wide executable.

Last: a 'pre-push' task is to be added that cleans the build folder,
runs phpunit, behat and phpCodeSniffer to make sure the project is
in optimal shape.

After cleaning up the project should these steps be added to the
contributor's guide.