Fix #583: Duplicate constants will override earlier definitions
Both the File and Interface reflectors made the list of constants
unique by assigning the name of the constant as array key and thus
overwriting earlier definitions.

Since it is assumed acceptable to have duplicate definitions that
are separated by a IF statement the code is to be altered so that
duplicates are allowed. This immediately also fixes the @ignore issue
that was mentioned in #583 as was observed during functional testing.