Fix #551: Remove POSIX collation element start
The License regular expression contains a character class containing
only a dot (`[.]`). In newer versions of POSIX this is recognized as the start
of a collating element. This causes issues in systems using a newer version of
POSIX as described in issue #551.

By replacing the character class containing the dot with an escaped dot we
circumvent this issue.

Thanks to @rvanvelzen for this solution.