Removed external class reflection
As discussed in PR #660 I have removed the support for Dynamic Reflection
of classes in the include path. @shrink encountered files that relied on
bootstrapping or changed in the include_path to work; and this threw a fatal
with phpDocumentor. And in turn this caused phpDocumentor to die fatally.

Given that this cannot be solved reliably have I chosen to remove the feature.