@boenrobot indicated that zend.script_encoding is used by PHP to know what the default input encoding is for PHP files and that this can be used as default.
As such I have changed the default to the zend.script_encoding setting (if set)
and if not set default to utf-8. This default is overridable in the
configuration or command line.

Another issue was the shut up operator; apparently there is a command that
disables the generation of PHP Warnings. I have altered the code to use that
and output the messages generated by libxml.