#697: Multiple templates in configuration only executes first one
Due to the way the config is parsed (and the lack of a Query language) was only
the first template picked. With these changes I have enabled a more robust way
of extracting the templates. However; the command handling should be refactored
to be more robust overall.

Included in this commit is the reformatting of the config template and the config
of phpDocumentor itself