#779: Removed copying of template's JS folder
In #779 it was reported that a generated template would throw an error
when used. This is caused because the template of a template has no
JS folder. This commit removes the offending action so that the error
is no longer reported as there is no JS to copy by default. 