Fix missing messages and re-add target parameter
Due to the refactoring in the previous commit was an error introduced
that caused all messages in the core plugin not to be shown. This
was caused because Zend\I18n reindexes the arrays for messages when
you have multiple sources. Since the original codes were purel numeric
they were re-ordered as different integers.

In addition to the above the target option was re-added to the parse command
as a simpler version.