Fix #802; structure.xml should contain full messages
Errors in the generated structure.xml did not contain a complete
error message, instead the error code was registered. To fix this
issue a new interface for writers was introduced to declare that
they should support translation of items and the XML Writer
should use that to translate the code into a message.