Create a series of requirements in Gherkin
phpDocumentor was missing a clear list of features, or requirements; to fill this gap I decided to introduce
more Behat feature files and populate them with automated tests along the way. The intended benefit is
that it becomes clear to a user which functionality is supported, for a contributor how the application
functions and for the project to track its goals and prevent regression issues.

The total effort of filling these is expected to take several weeks, if not several months, but will bring a
lot of stability and trust in the project.