Fixes #932: TODO tags did not show in marker report
The marker report with twig templates did not feature
TODO tags because these were never collected and
added to the marker index. XSL templates solved this
by using an, expensive, XPath query.

To solve this issue we have added a new compiler
pass that extracts TODO tags from all elements, adds
them to the marker listing and we have removed the
workaround from the XSL templates.