#1049: By reference was not passed from the Reflector to the Assembler
When an argument was declared as being 'by reference'
this was not registered on the Descriptor using the
Assemblers, and the XML writer did not set this in the
structure.xml.

During research it was also discovered that the
responsive template does not show calls by reference,
and probably the clean template not as well. This will
be addressed in a separate commit in the respective
repository.