Adding configuration for box-project.org
In order to reduce maintenance we want to build our phar files using @kherge's
box project. With this commit we add the configuration necessary to do so.
Unfortunately during tests it was unable to generate a phar file due to an
issue with open files.

I have opened issue https://github.com/kherge/Box/issues/64 to research the
cause of this issue so that we can generate a phar file at a later date.