Add BDD Scenarios for Inheritance
In an effort to fix all issues with inheritance in one go I have created
BDD tests to ensure all inheritance related activities work as expected.
This has resulted in more than 50 scenario's that are to be tested and
according to the latest run over half fail.

This provided a solid baseline to fix the issues related to inheritance.

These tests are related to issue #1031