Namespace/package index blocks show empty
On the frontpage there are blocks that contain Namespaces or Packages
so that you can easily go there. These blocks were incorrectly filled
due to an error with the population of the XML and the way the XSL
worked.

This commit fixes that.

Related to https://github.com/phpDocumentor/phpDocumentor2/issues/1158