Fixes #1229: @method tag in a trait causes Exception
In the TraitDescriptor we assembled a MethodDescriptor when an `@method`
tag is used. But instead of reading the Method Name we tried to get a
property name from the `@method` tag definition.

Because this method is not present an exception was generated. This
commit fixes the issue by using the correct method to get the method
name.