Fixes #1193: Package tags don't inherit to classes
@bskendig reported that a class without package tag did
not fall into the package that was set on the file in which
it was contained.

After research it appeared that although interfaces and traits
correctly inherited the package tag that classes tried to use the
`setPackage` method instead of the tags collection.

The setPackage method is 'reserved' to set the Package descriptor
on a descriptor and by not setting the 'package' element in the
tags collection are other parts of the application not aware that
the class falls within a specific package.

(for the interested reader: the package descriptor is set in a
compiler pass based on the package element in the tags collection)