#1232: Routing crashes on magic property in trait
Magic methods and properties are dynamically added by the TraitDescriptor. It is
these new Method and PropertyDescriptors that are fed into the routing system to
generate a link to them. Apparently we had missed setting a parent (to which class
a property or method is related) to these magic methods and properties.

By adding a parent during generation we get the desired behaviour again