#1211: Loading a single plugin is not possible
Due to a bug in Zend Config it is not possible to distinguish between
a single value and an array of values when using the same array keys.

There is a workaround for this bug by adding extra checks but this was
not added to the loading of plugins. This commit adds this workaround
so it becomes possible to load a single plugin.