#1248, #1212: Move configuration to application-level
It is not possible to define plugins using user-specific
configuration defined using the `-c` option because the
user-specific configuration was (re)loaded during individual
commands and thus after the plugins were loaded.

With this commit I have moved the configuration loading to
the application level and with it simplified the process and
enabling configuring application-specific configuration using
a user-specific configuration file.