#713: XSL Templates are not found in Windows
When running phpDocumentor from Phar the file paths are
interpreted as if on Linux; but there is an exception in
the Xsl writer where, if on windows, the path is generated
slightly different.

These two facts conflicted and resulted in XSL template
files not being found on windows when using Phar. This change
checks if the code is running as a Phar and does not apply
the exception if so.