#1087: Fully support `@see`
In our documentation we have included that `@see` should be able to cope with
URLs, references to FQSEN's but also that a function/method name without anything
in front of it may either refer to a method in the current class, or a function
in the current namespace.

This type of relation was not supported by phpDocumentor because of the decoupled
nature of its architecture. In order to properly support this I have introduced a
meta type called `@context` that is translated in the linked to either the current
class or the current namespace.

So for example:

    `@see myFunction()` on a method in a class either refers to a method of said
    class if it exists; or if it doesn't it refers to a function in the namespace
    of the class.