#1186: Move checking of writer requirements
Due to the way writer requirements were implemented did a run
always fail if one of the writers failed to meet their requirements
regardless whether it was used in the application.

To simplify the installation routine (by removing the ext/xsl composer
dependencies in templates) and to fail specific to a template (paving
the road to a conditional writer) we have moved the validation of
requirements to the point where a template is loaded.