#1239: Caching of project settings is broken
The caching mechanism aggresively tries to remove all garbage from
the caches. Unfortunately it proved to be a bit too aggresive because
it always removed the setting key in the cache.

That meant that if the settings were overridden with a value other than
the defaults; that the application assumed the settings were changed
and always did a full run.