Routing crashes in Twig
Apparently the twig extension always expects a method
`convertToRootPath` in the extension class; despite that the callback
points to the Route Renderer object.

I couldn't find the cause of this behaviour in time so as a workaround
I added a method `convertToRootPath` that proxies to the Route Renderer.

Perhaps we can find the cause and remove this proxy method in the
future.