#1213: Resolve @see and @link to Markdown link
In descriptions it is allowed to use the @see and @link tags; these tags
provide a link to a URL or other element in the project. In this commit
I have added a compiler pass that visits every description and replaces
the `@see` and `@link` tag with a Markdown-formatted link to that URL
or href.

I have chosen for a compiler pass because the transformer is the only
part that really knows the URLs for Descriptors and has a list of all
elements that are documented in the current project.