Adapt routing to resolve FQSENs
Currently routing works by resolving a Descriptor to a URL; but the XSL
document generation does not work with Descriptors but relies on strings
representing a FQSEN.

Instead of converting FQSENs into Descriptors in the XML/XSL writer I
have chosen to implement a Routing Rule that tries to detect if a string
is a FQSEN and if so, convert it to a Descriptor. This will then be
passed on to the other rules and follow the normal rules.