Refactored vendor-path determination
I have changed the way the vendor path is determined because it would
seem there are several minor issues with it,

- The JmsSerializerServiceProvider became dependent on phpDocumentor
- The vendor path is a value; instead of appending it to the constructor
  is was better suited to add a values parameter to the phpDocumentor
  application with which settings can be set from the invoking class.
- One of the paths pointed to the wrong folder
- Simplified the way DOMPdf is invoked due to the new determination

In addition I also changed the error message when the autoloader
cannot be found to be more descriptive and help users when they
encounter it. We have had two support questions in the past that could
have been prevented by making this message more descriptive.