Re-enable all release actions
After today's successful release using this playbook I can now
re-enable the commented actions with some certainty.

I have also added a `release_type` variable that needs to be set
using the command line. This may either be major, minor or patch
and indicates which type of release this is.