#40: Basic Implementation for Documentation Standards
In this commit I have added on the previous commit to create a
baseline way of dealign with Documentation Standard. phpDocumentor
now supports rulesets in the same way as PHP_CodeSniffer and as
such allows the custom creation of standards.

With this feature it will be possible to finetune and tweak how
your documentation is validated.