Add HHVM compatibility
In this commit we have made the changes that are necessary to run phpDocumentor
using HHVM. HHVM currently does not support the Phar::running() method, which
does also mean that unexpected side-effects may happen when running the PHAR file
from HHVM.

The most important side effects when running the Phar file is that HHVM will load
the DOMPDF extension, which normally fails under Phar, and that under windows paths
may not be converted correctly in some cases.

The latter is not important at this moment because HHVM does not run in Windows
anyway.