#1048: Create a JSONP writer
For phpDocumentor to create templates that consume content using XHR requests we
actually need to do JSONP requests to retrieve content. This is because docs
made with phpDocumentor are often read offline without the use of a web server.

The absence of a web browser has the disadvantage that XHR requests are not
possible with the single exception of JSONP requests because those are not
hindered by cross domain request restrictions.