#1452: Add an append action to FileIo Writer
For Angular we want to split the template up into partials; but because
it is impossible to have Angular load those on demand due to the Cross
Origin Request restrictions we have to have dynamic templates embedded
in the main index.html application template.

Because it is legit to append the <script> tags for the embedded partials
after the closing <html> tag we can separate the templates into independent
files and concatenate them together while generating the template.

The append action that is added in this commit does exactly that; append
content from one file onto a destination file (and create it if it doesn't
exist)