Only overwrite a configuration option if it contains a value
During testing we discovered that some configuration options became
unavailable due to the absence of this change. Perfectly good values
were overwritten with an empty value because the overriding configuration
did not override it.

To fix this I have made a change that will only copy a value onto the
configuration if it contains a value. This may have an adverse effect in
the future if we would want to disable a config value but we'll have
to check that out if it occurs.