Add support for multiple backends
I have refactored the Parser of phpDocumentor to achieve several goals:

1. Remove business logic from the Parse Command class so that it becomes easier to use
   the parser from another context.
2. Decouple the Caching from the Parser so that the caching layer becomes entirely
   optional.
3. Separate the parser coordination from the actual parsing by introducing 'backends';
   these can deal with the actual parsing of individual files and populating the
   ProjectDescriptor.
4. Separate logging from the business logic.

In order to do this the architecture has been changed to make the parser responsible
for collecting files; booting the backends with the configuration and emitting events
so that logging and caching can be event-driven instead of coupled inside the parser.

While parsing a file the parser will find a suitable backend and, if one is found,
pass the file to the backend for parsing and to populate the project descriptor.