Render templates
For phpDocumentor 3 we want to render templates using the new RenderActions
system that passes through the command handler. This will allow plugins to
be more effective and it fits the new Architecture a lot better.

In this commit I have created several Render Actions based on existing writers where
some will even use flysystem to write their results to.