GH #1339: Remove erroneous messages about duplicate package tags
@rscarson found the origin for this bug and by sharing his findings I was
able to implement a solution. The problem was caused because a package
tag was added before the DocBlock was analyzed, which meant that during
the analysis of the DocBlock the package tag would be added another time
because that method only adds tags and does not replace them.

By moving the defaultPackageName logic _below_ the analysis of the DocBlock
it would override, instead of add, the entire package tag.