XSL cannot find view locations
The XSL writer was unable to find the view locations because the
business logic for scanning the template folders was not implemented.

In this commit I have added a new FileRepository whose function is to
find the filenames for individual template files and return the locations
where templates may be found so that the Twig handler can re-use that
same logic.