Make links work
Because structure.xml has changed over time, the files were saved in the wrong place for some of the links to work.
The files are now moved to the classes directory.
This had implications for the navigation panel. This panel is used in index.html as well as in all the files that are now moved to classes.
Therefore the links would have a different starting point based on which file they are in.
To fix this, object-index.xsl is not written to a variable anymore, but is now included in chrome.xsl.
The starting point for the link is passed as a variable.
Also, in order for some links in the properties panel to work, the files should be generated based on the full_name of a class instead of the path of a file.
This caused changes in file.xsl, because the $path variable could no longer be passed. Instead of the file, the template now had to match the class/interface/trait and all references had to move one parent up (../)