Adds validating middleware to file construction.
This pr adds an other concept of validation without using an external validator library.
This way we don't have to convert your object stucture to an array. Which makes it easier
to validate new kind of rules.

The middleware implementation is just a way how things could work. Validation can easely added
as a post parse step using a command in the Cli run command class.