Start Refactoring to tighten DDD setup and implement Hexagonal Architecture
In order to make it easier to contribute to the project we need to make sure
that we make it as easy as possible. One of the ways to do this is to ensure
that we have an architecture that is well-understood and supported by
modern insights.

At current the DDD movement is gaining traction and with recent developments
has the hexagonal architecture style (or ports and adapters) become a main-
stream part of our community.

The advantage of adopting these is that they are pre-existing ideas and we
can stand on the shoulders of giants, like others.