Start separating layers and refactor results
The phpDocumentor 3 renderer needs to benefit from the concept
of an improving insight into the architecture of phpDocumentor 3.

For this purpose we formalize what information is passed to Actions
in the form of a rendering context and we abstract file system
interactions into a Domain concept of Artefacts and Assets.