Remove Renderer interface
Since we are going to refactor the entire way we do Actions and ActionHandlers
I have decided to remove the renderer interface since it causes a naming confusion
with the Renderer module.

The original intent was that the Renderer module takes care of the rendering of the
Documentation to Artefacts and that individual Renderers were consumed by Actions
to render just a single Artefact.

However; it would appear that the Renderer interface was discerned as the heart of the
Renderer module (which isn't odd to assume) and I have removed it in an attempt to
remove that confusion.