Add tests for CommandlineOptionsMiddleware and disable ContainerDefinitions and Action Handlers
In the upcoming fixes for the Renderers and Action handlers we will be doing a large amount of
changes. Covering these parts with tests now is unfeasable and would be counterproductive. Instead
we are excluding it from the code coverage at the moment so that the numbers are representative for
the amount of working code.

In addition I have also excluded the container definitions file because it is absurd to test if an
array contains all necessary items of this scope and magnitude (also changes a lot). The functioning
of this file should be inferred from working integration tests.