Moved commands and their handlers to the DomainModel
Insights that I have gained in the past months is that commands are expected
to be part of the domain model and not the application. This makes sense if
you consider that to run the functionality of the domain that the handlers
contain the glue, or business logic, to tie it together.