Removed the bin specifier from composer
phpDocumentor is not intended to be included in other applications using
the composer.json. There are too many dependencies that are shared by
other applications.

As such we will be actively discouraging this type of use, starting by
removing the line that will move the binary file into the vendor/bin folder.