If applied, this PR implements the following:
- Expands the basic *--ignore* examples for new users.
- Provides examples for single, and multiple directories.

I was trying to configure phpDocumentor to work with multiple top-level directories and found that the lack of basic --ignore syntax was a bit confusing. I had to look at stackoverflow to get help, which usually means that the core project documentation is missing some information. 

Please work with me to correct any project-specific style issues with my proposed changes. I felt it better to propose a starting point, from which we could collaborate to get it ready for inclusion.